+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix sticky table cell borders not rendering correctly when scrolling large tables in Trilium 0.47.
 - Fix attribute values not being separated correctly when there are more than two values displayed in a single table cell. Only the last two values would be separated by a space or line break instead of all values.
 - Fix `boolean` checkboxes being inverted (true as unchecked, false as checked).
+- Fix `prefix` and `suffix` settings not working with `boolean` checkboxes.
 
 ## 1.0.0 - 2021-04-17